+++ PatchNet/DP.h
@@ -13,26 +13,26 @@ using namespace std;
 class DP
 {
 public:
-	int _matrix[1000][1000];
-	int _matrix_min[1000][1000];
+	double _matrix[1000][1000];
+	double _matrix_min[1000][1000];
 	int _matrix_row[1000][1000]; //row of all
 	int _matrix_col[1000][1000]; //row of all
 
 	int _cols;//number of column or row
 	int _rows;;
 
-	int _mindistance;
-	int _maxdiatance;
+	double _mindistance;
+	double _maxdiatance;
 	vector<pair<int,int>> _path;
 
 	void initilize();
 	void printMatrix(int rows,int cols);
-	int MinMatrix(int i,int j);
-	int MaxMatrix(int i,int j);
+	double MinMatrix(int i,int j);
+	double MaxMatrix(int i,int j);
 	void performDP();
 
 public:
-	DP(int** distance,int rows,int cols);
+	DP(double** distance,int rows,int cols);
 	DP(void);
 	~DP(void);
 };
